add sweet alerts in multiple components

+++ libs/society-core/src/lib/components/config/plot-natures/plot-natures.component.ts
@@ -4,6 +4,7 @@ import { MyFormField, PlotNatureInterface } from '@society/shared/interface';
 import { SharedServicesDataModule } from '@society/shared/services/data';
 import { SharedServicesGlobalDataModule } from '@society/shared/services/global-data';
 import { PlotNaturesTableComponent } from './plot-natures-table/plot-natures-table.component';
+import Swal from "sweetalert2/dist/sweetalert2.js";
 
 @Component({
   selector: 'society-plot-natures',
@@ -109,27 +110,41 @@ export class PlotNaturesComponent implements OnInit {
     
     if(obj.num == '2')
     {
-      this.dataService
-      .deleteHttp(
-        this.pageFields,
-        this.formFields,
-        'society-api/PlotNature/savePlotNature'
-      )
-      .subscribe(
-        (response: any[]) => {
-          if(response[0].includes('Success') == true){
-            this.valid.apiInfoResponse('Record deleted successfully');
-            this.plotNatureTable.getPlotNature();
-            this.reset();
-          }else{
-            this.valid.apiErrorResponse(response[0]);
-          }
-        },
-        (error: any) => {
-          this.error = error;
-          this.valid.apiErrorResponse(this.error);
+      Swal.fire({
+        title: "Do you want to delete record?",
+        text: "",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes",
+        cancelButtonText: "No",
+      }).then((result: any) => {
+        if (result.value) {
+          this.dataService
+          .deleteHttp(
+            this.pageFields,
+            this.formFields,
+            'society-api/PlotNature/savePlotNature'
+          )
+          .subscribe(
+            (response: any[]) => {
+              if(response[0].includes('Success') == true){
+                this.valid.apiInfoResponse('Record deleted successfully');
+                this.plotNatureTable.getPlotNature();
+                this.reset();
+              }else{
+                this.valid.apiErrorResponse(response[0]);
+              }
+            },
+            (error: any) => {
+              this.error = error;
+              this.valid.apiErrorResponse(this.error);
+            }
+          );
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire("Cancelled", "", "error");
         }
-      );
+      });
+      
     }
   }
